add actions to main page
