+++ client/src/components/Main.jsx
@@ -1,7 +1,7 @@
 import axios from "axios";
 import { BASE_URL } from "../utils/constants";
 import { useEffect, useState } from "react";
-import ProductCart from "./ProductCart";
+import ProductCard from "./ProductCard";
 
 const Main = () => {
   const [products, setProducts] = useState([]);
@@ -26,7 +26,7 @@ const Main = () => {
   ) : (
     <div className="p-6 grid gap-6 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4">
       {products.map((pro) => (
-        <ProductCart key={pro._id} pro={pro} />
+        <ProductCard key={pro._id} pro={pro} />
       ))}
     </div>
   );
